+++ packages/cursorless-engine/src/spokenForms/defaultSpokenFormMap.types.ts
@@ -1,4 +1,5 @@
-import { SpokenFormMapKeyTypes, SpokenFormMappingType } from "./SpokenFormMap";
+import { SpokenFormMappingType } from "./SpokenFormMap";
+import { SpokenFormMapKeyTypes } from "./SpokenFormType";
 
 export type DefaultSpokenFormMapDefinition = {
   readonly [K in keyof SpokenFormMapKeyTypes]: Readonly<
